refactor: move multiindex handling to separate class
